refactored PotteryHTML and Inventory to fetch the database for pottery objects and list them.
