+++ index.html
@@ -8,6 +8,6 @@
   <body>
     <main class="container"></main>
     <article id="inventory"></article>
-    <script src="./scripts/main.js"></script>
+    <script type="module" src="./scripts/main.js"></script>
   </body>
 </html>
