+++ scripts/main.js
@@ -1,0 +1,4 @@
+import { getPottery } from "./partIII/PotteryDataProvider.js";
+import { PotteryInventory } from "./partII/PotteryInventory.js";
+
+getPottery().then(PotteryInventory);
