+++ scripts/partI/PotteryMaker.js
