+++ scripts/partII/PotteryHTML.js
@@ -3,9 +3,9 @@ export const PieceOfPottery = (potteryObj) => {
     <section id="piece">
         <h4>${potteryObj.shape}</h4>
         <ul>
-            <li>${potteryObj.height}</li>
-            <li>${potteryObj.weight}</li>
-            <li>$${potteryObj.price}</li>
+            <li>Height: ${potteryObj.height}</li>
+            <li>Weight: ${potteryObj.weight}</li>
+            <li>Price: $${potteryObj.price}</li>
 
         </ul>
         <button id="sellPottery--${potteryObj.id}">Sell</button>
